support mbz album detail view
